Rework Santa Tracker smoke tests
So they exercise
- all latests AGP versions >= 3.6
- both Kotlin and Java flavours of Santa Tracker

Before this commit, the only tested AGP version was 3.6.0-beta04
After this commit, 3.6.0-rc01, 4.0.0-alpha08, 4.0.0-<nightly> are tested

Tests on the Java flavour of Santa do pass with instant execution
